Couple small changes:
Added some notes on security to the call_handling
library.  Ultimately, I'd like to provide the
gem w/the ability to verify secure twiml requests.

Also, tweaked Account along the same lines,
basically returning nil when the twiml request
doesn't contain a sid.

Important note for the latter, I'm handling
whether or not we have the correct for the calling
app (but this is a TODO: to be changed).